[style]: reorganize code
